Fix how pbft deal with tx conflicts + other stuff (#1064)

+++ src/consensus/pbft/libbyz/request.h
@@ -23,9 +23,8 @@ struct Request_rep : public Message_rep
                  // expects to receive a full reply
                  // (if negative, it means all replicas).
   short command_size;
-  int cid; // unique id of client who sends the request
-  uint64_t
-    uid; // a way for a client_proxy to group requests from different users
+  short cid; // unique id of client who sends the request
+  short uid; // a way for a client_proxy to group requests from different users
   Request_id rid; // unique request identifier
   // Followed a command which is "command_size" bytes long and an
   // authenticator.
